Implemented CNN CSV data input, fixed np error for float classes
The CNN data set option was introduced to receive feature vector CSVs as input. Theoretically, this option could be used to accept any generic CSVs.

The test data option, -x, seems to be deprecated.

A numpy list/array error was fixed for all float classes.

This code was used for the DEMUD-0 experiment.